+++ data/src/main/java/com/ssing/data/lesson/remote/dto/response/InstructorLessonDetailOngoingResponse.kt
@@ -40,8 +40,8 @@ internal data class InstructorLessonDetailOngoingResponse(
 
     @Serializable
     internal data class MatchingRequest(
-        @SerialName("matchingRequestId") val matchingRequestId: Int,
-        @SerialName("representativeMemberId") val representativeMemberId: Int,
+        @SerialName("matchingRequestId") val matchingRequestId: Long,
+        @SerialName("representativeMemberId") val representativeMemberId: Long,
         @SerialName("representativeMemberName") val representativeMemberName: String,
         @SerialName("headcount") val headcount: Int,
         @SerialName("teamLessonPrice") val teamLessonPrice: Int,
@@ -50,7 +50,7 @@ internal data class InstructorLessonDetailOngoingResponse(
 
     @Serializable
     internal data class Participant(
-        @SerialName("participantId") val participantId: Int,
+        @SerialName("participantId") val participantId: Long,
         @SerialName("gender") val gender: String,
         @SerialName("age") val age: Int,
     )
